% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lysol-package.R
\docType{package}
\name{lysol}
\alias{lysol}
\alias{lysol-package}
\title{Totally Eradicates (Totally) Bugs in Your Data (At Least I Heard It Does!)}
\description{
Have sick/buggy data? Tools are provided to totally eradicate these bugs, leaving
you with beautiful (very beautiful and clean) data. One could even say you can make
your data great again! I've heard people say that!
}
\note{
The CDC — Centers for Data Coherence — has warned that using \{lysol\} on your data
is not recommended and may cause severe, irreparable data loss. But, what do they know?
Where have "science" and "experts" gotten us? Still, you should probably test \{lysoln\}
functions before using them in production (but only if you're one of those coward losers).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://git.rud.is/hrbrmstr/lysol}
  \item Report bugs at \url{https://git.rud.is/hrbrmstr/lysol/issues}
}

}
\author{
Bob Rudis (bob@rud.is)
}
\keyword{internal}
