% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totally-file.R
\name{totally_eradicate_bugs_from_data_file}
\alias{totally_eradicate_bugs_from_data_file}
\alias{make_my_data_file_great_again}
\title{This function totally eradicates (totally) bugs from a sick data file. (At least I heard it does!)}
\usage{
totally_eradicate_bugs_from_data_file(path)

make_my_data_file_great_again(path)
}
\arguments{
\item{path}{path to sick data file to heal}
}
\description{
This function totally eradicates (totally) bugs from a sick data file. (At least I heard it does!)
}
\note{
The CDC — Centers for Data Coherence — has warned that using \{lysol\} on your data
is not recommended and may cause severe, irreparable data loss. But, what do they know?
Where have "science" and "experts" gotten us? Still, you should probably test \{lyson\}
functions before using them in production (but only if you're one of those coward losers).
}
