% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/totally-object.R
\name{totally_eradicate_bugs_from_this}
\alias{totally_eradicate_bugs_from_this}
\alias{make_this_data_great_again}
\title{This function totally eradicates (totally) bugs from a sick objects.  (At least I heard it does!)}
\usage{
totally_eradicate_bugs_from_this(x)

make_this_data_great_again(x)
}
\arguments{
\item{x}{R object to cleanse}
}
\value{
A totally cleansed object.
}
\description{
This function totally eradicates (totally) bugs from a sick objects.  (At least I heard it does!)
}
\note{
The CDC — Centers for Data Coherence — has warned that using \{lysol\} on your data
is not recommended and may cause severe, irreparable data loss. But, what do they know?
Where have "science" and "experts" gotten us? Still, you should probably test \{lysoln\}
functions before using them in production (but only if you're one of those coward losers).
}
